requirements.txt
web.py
python-daemon
